Re-enable global cache config
