+++ packages/core/src/Cache/StorageCacheCommon.ts
@@ -1,6 +1,7 @@
 // Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
 // SPDX-License-Identifier: Apache-2.0
 
+import { Amplify } from '../singleton';
 import { ConsoleLogger as Logger } from '../Logger';
 import { KeyValueStorageInterface } from '../types';
 import { currentSizeKey, defaultConfig } from './constants';
@@ -30,8 +31,7 @@ export abstract class StorageCacheCommon {
 		config?: CacheConfig;
 		keyValueStorage: KeyValueStorageInterface;
 	}) {
-		// const globalCacheConfig = Amplify.getConfig().Cache ?? {};
-		const globalCacheConfig = {};
+		const globalCacheConfig = Amplify.getConfig().Cache ?? {};
 
 		this.config = {
 			...defaultConfig,
