+++ packages/core/src/Cache/types/cache.ts
@@ -1,6 +1,9 @@
 // Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
 // SPDX-License-Identifier: Apache-2.0
 
+import { CacheConfig } from '../../singleton/Cache/types';
+export { CacheConfig };
+
 /**
  * Cache Interface
  */
@@ -28,31 +31,6 @@ export interface Cache {
 
 	/** change current configuration */
 	configure(config: CacheConfig): CacheConfig;
-}
-
-/**
- * Cache instance options
- */
-export interface CacheConfig {
-	/** Prepend to key to avoid conflicts */
-	keyPrefix: string;
-
-	/** Cache capacity, in bytes */
-	capacityInBytes: number;
-
-	/** Max size of one item */
-	itemMaxSize: number;
-
-	/** Time to live, in milliseconds */
-	defaultTTL: number;
-
-	/** Warn when over threshold percentage of capacity, maximum 1 */
-	warningThreshold: number;
-
-	/** default priority number put on cached items */
-	defaultPriority: number;
-
-	storage?: Storage;
 }
 
 export interface CacheItem {
